Move delete/confirm utility functions
So they can be reused by non-admin views
