+++ tests/unit/test_routes.py
@@ -144,37 +144,37 @@ def test_routes(warehouse):
         ),
         pretend.call(
             "manage.project.settings",
-            "/manage/project/{name}/settings/",
+            "/manage/project/{project_name}/settings/",
             factory="warehouse.packaging.models:ProjectFactory",
-            traverse="/{name}",
+            traverse="/{project_name}",
             domain=warehouse,
         ),
         pretend.call(
             "manage.project.releases",
-            "/manage/project/{name}/releases/",
+            "/manage/project/{project_name}/releases/",
             factory="warehouse.packaging.models:ProjectFactory",
-            traverse="/{name}",
+            traverse="/{project_name}",
             domain=warehouse,
         ),
         pretend.call(
             "manage.project.roles",
-            "/manage/project/{name}/collaboration/",
+            "/manage/project/{project_name}/collaboration/",
             factory="warehouse.packaging.models:ProjectFactory",
-            traverse="/{name}",
+            traverse="/{project_name}",
             domain=warehouse,
         ),
         pretend.call(
             "manage.project.change_role",
-            "/manage/project/{name}/collaboration/change/",
+            "/manage/project/{project_name}/collaboration/change/",
             factory="warehouse.packaging.models:ProjectFactory",
-            traverse="/{name}",
+            traverse="/{project_name}",
             domain=warehouse,
         ),
         pretend.call(
             "manage.project.delete_role",
-            "/manage/project/{name}/collaboration/delete/",
+            "/manage/project/{project_name}/collaboration/delete/",
             factory="warehouse.packaging.models:ProjectFactory",
-            traverse="/{name}",
+            traverse="/{project_name}",
             domain=warehouse,
         ),
         pretend.call(
