+++ warehouse/routes.py
@@ -110,37 +110,37 @@ def includeme(config):
     config.add_route("manage.projects", "/manage/projects/", domain=warehouse)
     config.add_route(
         "manage.project.settings",
-        "/manage/project/{name}/settings/",
+        "/manage/project/{project_name}/settings/",
         factory="warehouse.packaging.models:ProjectFactory",
-        traverse="/{name}",
+        traverse="/{project_name}",
         domain=warehouse,
     )
     config.add_route(
         "manage.project.releases",
-        "/manage/project/{name}/releases/",
+        "/manage/project/{project_name}/releases/",
         factory="warehouse.packaging.models:ProjectFactory",
-        traverse="/{name}",
+        traverse="/{project_name}",
         domain=warehouse,
     )
     config.add_route(
         "manage.project.roles",
-        "/manage/project/{name}/collaboration/",
+        "/manage/project/{project_name}/collaboration/",
         factory="warehouse.packaging.models:ProjectFactory",
-        traverse="/{name}",
+        traverse="/{project_name}",
         domain=warehouse,
     )
     config.add_route(
         "manage.project.change_role",
-        "/manage/project/{name}/collaboration/change/",
+        "/manage/project/{project_name}/collaboration/change/",
         factory="warehouse.packaging.models:ProjectFactory",
-        traverse="/{name}",
+        traverse="/{project_name}",
         domain=warehouse,
     )
     config.add_route(
         "manage.project.delete_role",
-        "/manage/project/{name}/collaboration/delete/",
+        "/manage/project/{project_name}/collaboration/delete/",
         factory="warehouse.packaging.models:ProjectFactory",
-        traverse="/{name}",
+        traverse="/{project_name}",
         domain=warehouse,
     )
