+++ warehouse/templates/manage/roles.html
@@ -61,7 +61,7 @@
         {% if role.user == request.user %}
           {{ role.role_name }}
         {% else %}
-          <form class="changeRole" method="POST" action="{{ request.route_path('manage.project.change_role', name=project.name) }}">
+          <form class="changeRole" method="POST" action="{{ request.route_path('manage.project.change_role', project_name=project.name) }}">
             <input name="csrf_token" type="hidden" value="{{ request.session.get_csrf_token() }}">
             {% for role in roles %}
               <input type="hidden" name="role_id" value="{{ role.id }}">
@@ -85,7 +85,7 @@
         {% endif %}
         </td>
         <td>
-          <form method="POST" action="{{ request.route_path('manage.project.delete_role', name=project.name) }}">
+          <form method="POST" action="{{ request.route_path('manage.project.delete_role', project_name=project.name) }}">
             <input name="csrf_token" type="hidden" value="{{ request.session.get_csrf_token() }}">
             {#
             The following three lines are added to handle multiple roles for a
