+++ warehouse/templates/packaging/detail.html
@@ -112,7 +112,6 @@
   </div>
 </section>
 
-
 <section>
   <div class="site-container">
     <div class="vertical-tabs">
